Added a CircleCI configuration file to get continous deployments to AWS
This CircleCI configuration is modified from the config I use on my personal sites jonrh.is and bespokedashboards.com. This config alone is not enough. We will have to define the $CLOUDFRONTDISTRIBUTIONID environment variable on the CircleCI website for the project. Furthermore we will have to create the apropriet AWS access roles so CircleCI has permissions to do what it needs to do.